refactor(schedule): extract schedule utils and types to shared modules
