+++ apps/cugetreg/src/routes/schedule/+page.svelte
@@ -1,6 +1,8 @@
 <script lang="ts">
+  import { browser } from '$app/environment';
   import AppSidebar from '$lib/components/app-sidebar.svelte';
   import SelectedCourse from '$lib/components/selected-course.svelte';
+  import { getColumnFromDay } from '$lib/mapper';
   import {
     getSemesterShortOptions,
     getYearOptions,
@@ -11,9 +13,18 @@
     getUserCartStore,
     useCartActions,
   } from '$lib/stores/user-cart';
+  import {
+    calculateCredit,
+    createElementScreenshot,
+    getExamData,
+    getExamDateOrder,
+    getViewCourseData,
+    isConflicted,
+    isExamConflicted,
+    parsePeriodTime,
+  } from '$lib/utils/schedule';
 
   import { BookMarked, Loader2, Menu } from '@lucide/svelte';
-  import html2canvas from 'html2canvas-pro';
   import {
     ChevronLeft,
     ChevronRight,
@@ -50,16 +61,14 @@
   import * as Sidebar from '@cugetreg/ui/organisms/sidebar';
   import { ViewCourse } from '@cugetreg/ui/organisms/view-course';
   import {
-    discardTime,
     formatDate,
     formatExamColumn,
     formatExamTime,
   } from '@cugetreg/utils';
   import type { ColorVariant, Day } from '@cugetreg/utils/types';
-  import type {
-    CartItemDetail,
-    Period,
-  } from '@cugetreg/zod-schemas/carts-response';
+  import type { CartItemDetail } from '@cugetreg/zod-schemas/carts-response';
+
+  // ================ HOOKS ================
 
   const userCart = getUserCartStore();
   const {
@@ -74,94 +83,13 @@
     updateCartMeta,
   } = useCartActions();
 
-  // TODO: Move this somewhere else
-  function parsePeriodTime(periodTime: string): number {
-    const parts = periodTime.split(':');
-    const hour = Number(parts[0]);
-    const minute = Number(parts[1]);
+  const cartPromise = getContext<CartPromise>(CART_PROMISE_KEY);
 
-    return hour + minute / 60;
-  }
-
-  function getColumnFromDay(day: Day): number {
-    switch (day) {
-      case 'MO':
-        return 0;
-      case 'TU':
-        return 1;
-      case 'WE':
-        return 2;
-      case 'TH':
-        return 3;
-      case 'FR':
-        return 4;
-      case 'SA':
-        return 5;
-      case 'SU':
-        return 6;
-    }
-  }
-
-  function isConflicted(courseNo: string, period: Period): boolean {
-    for (const other of $userCart.currentCart.items) {
-      if (other.hidden || other.courseNo === courseNo) continue;
-
-      const otherSection = other.sections.find(
-        (sec) => sec.sectionNo === other.sectionNo,
-      );
-
-      const periodStartTime = parsePeriodTime(period.periodStart);
-      const periodEndTime = parsePeriodTime(period.periodEnd);
-
-      if (isNaN(periodStartTime) || isNaN(periodEndTime)) continue;
-
-      for (const otherPeriod of otherSection?.classes ?? []) {
-        const otherPeriodStartTime = parsePeriodTime(otherPeriod.periodStart);
-        const otherPeriodEndTime = parsePeriodTime(otherPeriod.periodEnd);
-
-        if (isNaN(otherPeriodStartTime) || isNaN(otherPeriodEndTime)) continue;
-
-        if (period.dayOfWeek !== otherPeriod.dayOfWeek) continue;
-        if (
-          periodStartTime < otherPeriodEndTime &&
-          otherPeriodStartTime < periodEndTime
-        ) {
-          return true;
-        }
-      }
-    }
-    return false;
-  }
-
-  async function screenshotTimetable() {
-    if (!timetableDiv) return;
-
-    const canvas = await html2canvas(timetableDiv, {
-      backgroundColor: null,
-      logging: false,
-      useCORS: true,
-      scale: 3,
-    });
-
-    const screenshot = canvas.toDataURL('image/jpeg');
-
-    const link = document.createElement('a');
-    link.href = screenshot;
-    link.download = `${$userCart.currentCart.name}_timetable.jpg`;
-    document.body.appendChild(link);
-    link.click();
-    document.body.removeChild(link);
-  }
-
-  function handlePublicToggle(isPublic: boolean) {
-    updateCartMeta({
-      visible: isPublic ? 'PUB' : 'PVT',
-    });
-  }
+  // ================ STATES ================
 
   let showExamSchedule = $state<'List' | 'Schedule'>('Schedule');
 
-  let timetableDiv = $state<HTMLElement | null>(null);
+  let timetableDiv = $state<HTMLElement | undefined>();
   let innerWidth = $state(1024);
 
   let showRenameScheduleModal = $state(false);
@@ -169,7 +97,7 @@
   let showDeleteScheduleModal = $state(false);
   let showViewCourseModal = $state(false);
 
-  let selectedCartItemId = $state<string | null>(null);
+  let selectedCartItemId = $state<string | undefined>(undefined);
 
   let scheduleTableRef = $state<HTMLElement | null>(null);
   let midtermTableRef = $state<HTMLElement | null>(null);
@@ -179,172 +107,27 @@
   let openPanel = $state<'sidebar' | 'selected_only' | null>(null);
   let activePanel = $state<'sidebar' | 'selected_only' | null>(null);
 
-  const viewCourseData = $derived.by(() => {
-    if (!selectedCartItemId) return null;
-
-    const selectedCartItem = $userCart.currentCart?.items.find(
-      (item) => item.id === selectedCartItemId,
-    );
-    if (!selectedCartItem) return null;
-
-    const midterm = $userCart.exams?.find(
-      (e) => e.cartItemId === selectedCartItem?.id && e.type === 'MIDTERM',
-    );
-    const final = $userCart.exams?.find(
-      (e) => e.cartItemId === selectedCartItem?.id && e.type === 'FINAL',
-    );
-
-    const data: ViewCourseData = {
-      itemId: selectedCartItem.id,
-      courseNo: selectedCartItem.courseNo,
-      abbrName: selectedCartItem.course.abbrName,
-      courseNameTh: selectedCartItem.course.courseNameTh,
-      courseNameEn: selectedCartItem.course.courseNameEn,
-      credit: selectedCartItem.course.credit,
-      sections: selectedCartItem.sections.map((sec) => ({
-        sectionNo: sec.sectionNo,
-        closed: sec.closed,
-        regis: sec.regis,
-        max: sec.max,
-        classes: sec.classes.map((cls) => ({
-          type: cls.type,
-          dayOfWeek: cls.dayOfWeek,
-          periodStart: cls.periodStart,
-          periodEnd: cls.periodEnd,
-          building: cls.building,
-          room: cls.room,
-          professors: cls.professors,
-        })),
-      })),
-      selectedSectionNo: selectedCartItem.sectionNo,
-      color: (selectedCartItem.color as ColorVariant) ?? 'primary',
-      midterm: midterm
-        ? `${formatDate(new Date(midterm.start))} ${formatExamTime(new Date(midterm.start), (new Date(midterm.end).getTime() - new Date(midterm.start).getTime()) / (1000 * 60 * 60))}`
-        : undefined,
-      final: final
-        ? `${formatDate(new Date(final.start))} ${formatExamTime(new Date(final.start), (new Date(final.end).getTime() - new Date(final.start).getTime()) / (1000 * 60 * 60))}`
-        : undefined,
-      isHidden: selectedCartItem.hidden,
-    };
-
-    return data;
-  });
-
-  type LocalExamData = {
-    abbrName: string;
-    cartItemId: string;
-    courseNo: string;
-    name: string;
-    colorVariant: ColorVariant;
-    start: Date | null;
-    end: Date | null;
-    duration: number; // calculated here
-  };
-
-  const examsData = $derived.by(() => {
-    let midterms: Record<number, LocalExamData[]> = {};
-    let finals: Record<number, LocalExamData[]> = {};
-
-    if ($userCart && $userCart.exams) {
-      for (const exam of $userCart.exams) {
-        const course = $userCart.currentCart?.items.find(
-          (item) => item.id === exam.cartItemId,
-        );
-        if (!course || course.hidden) continue;
-
-        let dateVal = 0;
-        let start: Date | null = null;
-        let end: Date | null = null;
-        let duration = 0;
-
-        if (exam.start && exam.end) {
-          start = new Date(exam.start);
-          end = new Date(exam.end);
-          dateVal = discardTime(start.getTime());
-          duration = (end.getTime() - start.getTime()) / (1000 * 60 * 60);
-        }
-
-        const data: LocalExamData = {
-          cartItemId: exam.cartItemId,
-          courseNo: exam.courseNo,
-          name:
-            course.course.courseNameEn ||
-            course.course.courseNameTh ||
-            exam.courseNo,
-          abbrName: course.course.abbrName,
-          colorVariant: (course.color ?? 'primary') as ColorVariant,
-          start,
-          end,
-          duration,
-        };
-
-        if (exam.type === 'MIDTERM') {
-          if (!midterms[dateVal]) midterms[dateVal] = [];
-          midterms[dateVal].push(data);
-        } else if (exam.type === 'FINAL') {
-          if (!finals[dateVal]) finals[dateVal] = [];
-          finals[dateVal].push(data);
-        }
-      }
-    }
-
-    return { midterms, finals };
-  });
-
-  const examSort = (a: string, b: string) => {
-    const numA = Number(a);
-    const numB = Number(b);
-
-    if (numA === 0) return 1;
-    else if (numB === 0) return -1;
-    else return numA - numB;
-  };
-
-  const examDateOrder = $derived.by(() => {
-    let midterms: number[] = Object.keys(examsData.midterms)
-      .sort(examSort)
-      .map((x) => Number(x));
-    let finals: number[] = Object.keys(examsData.finals)
-      .sort(examSort)
-      .map((x) => Number(x));
-
-    return { midterms, finals };
-  });
-
-  function isExamConflicted(exam: LocalExamData, others?: LocalExamData[]) {
-    if (!exam.start || !exam.end) return false;
-    if (!others) return false;
-
-    const s1 = exam.start.getTime();
-    const e1 = exam.end.getTime();
-
-    for (const other of others) {
-      if (other.cartItemId === exam.cartItemId) continue;
-      if (!other.start || !other.end) continue;
-
-      const s2 = other.start.getTime();
-      const e2 = other.end.getTime();
-
-      if (s1 < e2 && s2 < e1) return true;
-    }
-
-    return false;
-  }
-
-  const totalCredit = $derived(
-    $userCart.currentCart?.items.reduce(
-      (acc, item) => acc + (item.hidden ? 0 : Number(item.course.credit)),
-      0,
-    ) ?? 0,
-  );
-
-  const publicCartURL = $derived(
-    `${window.location.host}/schedule/${$userCart.currentCart.id}`,
-  );
-
-  const cartPromise = getContext<CartPromise>(CART_PROMISE_KEY);
-
   let previousCartId = $state($userCart.currentCartId);
+
+  // ================ DERIVED ================
+
+  const viewCourseData = $derived(
+    getViewCourseData(
+      selectedCartItemId,
+      $userCart.currentCart,
+      $userCart.exams,
+    ),
+  );
+
+  const examsData = $derived(
+    getExamData($userCart.currentCart, $userCart.exams),
+  );
+  const examDateOrder = $derived(getExamDateOrder(examsData));
+  const totalCredit = $derived(calculateCredit($userCart.currentCart.items));
+  const publicCartURL = $derived.by(() => {
+    if (!browser) return '';
+    return `${window.location.host}/schedule/${$userCart.currentCart.id}`;
+  });
 
   $effect(() => {
     if ($userCart.currentCartId && $userCart.currentCartId !== previousCartId) {
@@ -352,6 +135,19 @@
       switchCart($userCart.currentCartId);
     }
   });
+
+  function handleTimetableScreenshot() {
+    createElementScreenshot(
+      `${$userCart.currentCart.name}_timetable`,
+      timetableDiv,
+    );
+  }
+
+  function handlePublicToggle(isPublic: boolean) {
+    updateCartMeta({
+      visible: isPublic ? 'PUB' : 'PVT',
+    });
+  }
 </script>
 
 <svelte:window bind:innerWidth />
@@ -621,7 +417,7 @@
               {#if innerWidth < 1024}
                 <Button
                   class="m-0 flex items-center gap-1 border border-gray-200 bg-white"
-                  onclick={screenshotTimetable}
+                  onclick={handleTimetableScreenshot}
                 >
                   <Download
                     size={20}
@@ -659,7 +455,7 @@
                 <Share2 />
               </IconButton>
               {#if innerWidth >= 1024}
-                <Button class="m-0" onclick={screenshotTimetable}>
+                <Button class="m-0" onclick={handleTimetableScreenshot}>
                   บันทึกเป็นภาพ
                 </Button>
               {/if}
@@ -874,7 +670,11 @@
     {#each section?.classes ?? [] as period, i (i)}
       {@const startTime = parsePeriodTime(period.periodStart)}
       {@const endTime = parsePeriodTime(period.periodEnd)}
-      {@const color = isConflicted(courseNo, period)
+      {@const color = isConflicted(
+        courseNo,
+        period,
+        $userCart.currentCart.items,
+      )
         ? 'conflict'
         : ((course.color as ColorVariant) ?? 'primary')}
