+++ packages/ui/src/lib/components/organisms/view-course/view-course.svelte
@@ -13,39 +13,7 @@
 	import { courseColorVariants } from '@cugetreg/utils/constants';
 	import { type ColorVariant } from '@cugetreg/utils/types';
 
-	export interface ViewCourseSectionClass {
-		type: string;
-		dayOfWeek: string;
-		periodStart: string;
-		periodEnd: string;
-		building: string | null;
-		room: string | null;
-		professors: string[];
-	}
-
-	export interface ViewCourseSection {
-		sectionNo: number;
-		closed: boolean;
-		regis: number;
-		max: number;
-		classes: ViewCourseSectionClass[];
-	}
-
-	export interface ViewCourseData {
-		itemId: string;
-		courseNo: string;
-		abbrName: string;
-		courseNameTh: string;
-		courseNameEn: string;
-		credit: string | number;
-		genEdType?: string;
-		sections: ViewCourseSection[];
-		selectedSectionNo: number;
-		color: ColorVariant;
-		midterm?: string;
-		final?: string;
-		isHidden: boolean;
-	}
+	import type { ViewCourseData } from '.';
 
 	interface ViewCourseProps {
 		data: ViewCourseData | null;
